More write reorganization
