+++ crates/polars-io/src/csv/read/utils.rs
@@ -14,7 +14,7 @@ use polars_utils::slice::GetSaferUnchecked;
 use crate::csv::read::parser::{is_comment_line, next_line_position_naive};
 #[cfg(any(feature = "decompress", feature = "decompress-fast"))]
 use crate::csv::read::parser::{next_line_position, skip_bom, skip_line_ending, SplitLines};
-use crate::csv::splitfields::SplitFields;
+use crate::csv::read::splitfields::SplitFields;
 use crate::csv::CsvEncoding;
 use crate::mmap::ReaderBytes;
 use crate::prelude::{CommentPrefix, NullValues};
