+++ crates/polars-io/src/csv/write/mod.rs
@@ -1,9 +1,10 @@
+mod write_impl;
+
 use std::num::NonZeroUsize;
 
 use polars_core::POOL;
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
-mod write_impl;
 pub use write_impl::SerializeOptions;
 
 use super::*;
